Move test to more appropriate location
