+++ api/src/test/java/org/snomed/snap2snomed/integration/repository/PerMethodResourceTestsIT.java
@@ -42,19 +42,6 @@ public class PerMethodResourceTestsIT extends IntegrationTestBase {
     restClient.createOrUpdateUser(PROJECT_USER, "Bob2", "Bobby2", "Bob2", "u3@csiro.au");
   }
 
-    @Test
-  public void givenAdminUser_whenNotProjectMember_thenShouldNotSeeImportedCodeSystem() throws Exception {
-    long projectId = restClient.createProject("ProjectDemo", "Demo Project", Set.of(DEFAULT_TEST_USER_SUBJECT, PROJECT_USER), Set.of(), Set.of());
-    final String codeSetName = "AAA semicolon - defaultuser - projecttest";
-    long codesetId = immportCodeSetForUser(DEFAULT_TEST_USER_SUBJECT, codeSetName, "1.0", 0, 2, true, ";", new ClassPathResource("AAA-semi.csv").getFile(), "text/tsv");
-
-    restClient.createMap("Testing Map Version", "http://snomed.info/sct/32506021000036107/version/20210531",
-            "http://map.test.toscope", projectId, codesetId);
-
-    restClient.givenUser(DEFAULT_TEST_ADMIN_USER_SUBJECT).get("/importedCodeSets")
-              .then().statusCode(200).body("content", not(hasItem(hasEntry("name", codeSetName))));
-  }
-
   /**
    * Tests that users can see an imported codeset who is associated with a project that has the importedcodset as a base
    */
